+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod disjoint_set;
+mod disjoint_set;
 pub mod export;
 pub mod pack;
 pub mod place;
