+++ src/texture/mod.rs
@@ -3,10 +3,12 @@ use std::{
     sync::mpsc,
 };
 
-use image::{DynamicImage, GenericImageView, ImageBuffer, ImageReader};
+use image::{DynamicImage, GenericImageView, ImageBuffer};
 use rayon::prelude::*;
-use stretto::Cache;
-use sys_info::mem_info;
+use utils::{calc_bbox, uv_to_pixel_coords};
+
+pub mod cache;
+mod utils;
 
 #[derive(Debug, Clone)]
 pub struct DownsampleFactor(f32);
@@ -22,103 +24,6 @@ impl DownsampleFactor {
 
     pub fn value(&self) -> f32 {
         self.0
-    }
-}
-
-// Cache for storing the only size of the image
-pub struct TextureSizeCache {
-    cache: Cache<PathBuf, (u32, u32)>,
-}
-
-impl TextureSizeCache {
-    pub fn new() -> Self {
-        TextureSizeCache {
-            cache: Cache::new(1_000_000, 1_000_000).unwrap(),
-        }
-    }
-
-    pub fn get_or_insert(&self, image_path: &PathBuf) -> (u32, u32) {
-        match self.cache.get(image_path) {
-            Some(size) => *size.value(),
-            None => {
-                let size = get_image_size(image_path).unwrap();
-                // Since it only retains the size of the texture, set the cost to 1 for everything.
-                let cost = 1;
-                self.cache.insert(image_path.to_path_buf(), size, cost);
-                self.cache.wait().unwrap();
-
-                size
-            }
-        }
-    }
-}
-
-impl Default for TextureSizeCache {
-    fn default() -> Self {
-        TextureSizeCache::new()
-    }
-}
-
-impl Drop for TextureSizeCache {
-    fn drop(&mut self) {
-        self.cache.close().unwrap();
-    }
-}
-
-// Cache for storing the image
-pub struct TextureCache {
-    cache: Cache<PathBuf, DynamicImage>,
-}
-
-impl TextureCache {
-    pub fn new(capacity: usize) -> Self {
-        let default_capacity = get_cache_size().unwrap();
-        if capacity == 0 {
-            TextureCache {
-                cache: Cache::new(default_capacity, 2_000_000_000).unwrap(),
-            }
-        } else {
-            TextureCache {
-                cache: Cache::new(capacity, 2_000_000_000).unwrap(),
-            }
-        }
-    }
-
-    pub fn get_image(&self, path: &PathBuf) -> DynamicImage {
-        match self.cache.get(path) {
-            Some(image) => image.value().clone(),
-            None => {
-                let image = image::open(path).expect("Failed to open image file");
-                let cost = image.width() * image.height() * image.color().bytes_per_pixel() as u32;
-                self.cache
-                    .insert(path.to_path_buf(), image.clone(), cost as i64);
-                self.cache.wait().unwrap();
-
-                image
-            }
-        }
-    }
-}
-
-fn get_cache_size() -> Result<usize, String> {
-    const MIN_CACHE_SIZE: usize = 100 * 1024 * 1024; // 100MB
-    const MAX_CACHE_SIZE: usize = 2 * 1024 * 1024 * 1024; // 2GB
-
-    match mem_info() {
-        Ok(mem) => {
-            let total_memory = mem.total as usize * 1024;
-            // 15% of total memory
-            let cache_size = (total_memory as f64 * 0.15) as usize;
-            Ok(cache_size.clamp(MIN_CACHE_SIZE, MAX_CACHE_SIZE))
-        }
-
-        Err(e) => Err(format!("Failed to retrieve memory information.: {}", e)),
-    }
-}
-
-impl Drop for TextureCache {
-    fn drop(&mut self) {
-        self.cache.close().unwrap();
     }
 }
 
@@ -293,33 +198,4 @@ fn is_point_inside_polygon(test_point: (f64, f64), polygon: &[(f64, f64)]) -> bo
     }
 
     is_inside
-}
-
-// utils
-
-fn get_image_size<P: AsRef<Path>>(file_path: P) -> Result<(u32, u32), image::ImageError> {
-    let reader = ImageReader::open(file_path)?;
-    let dimensions = reader.into_dimensions()?;
-    Ok(dimensions)
-}
-
-fn uv_to_pixel_coords(uv_coords: &[(f64, f64)], width: u32, height: u32) -> Vec<(u32, u32)> {
-    uv_coords
-        .iter()
-        .map(|(u, v)| {
-            (
-                (u.clamp(0.0, 1.0) * width as f64).min(width as f64 - 1.0) as u32,
-                ((1.0 - v.clamp(0.0, 1.0)) * height as f64).min(height as f64 - 1.0) as u32,
-            )
-        })
-        .collect()
-}
-
-fn calc_bbox(pixel_coords: &[(u32, u32)]) -> (u32, u32, u32, u32) {
-    pixel_coords.iter().fold(
-        (u32::MAX, u32::MAX, 0, 0),
-        |(min_x, min_y, max_x, max_y), (x, y)| {
-            (min_x.min(*x), min_y.min(*y), max_x.max(*x), max_y.max(*y))
-        },
-    )
 }
